Move AllSimplePaths to reachability
